backend/requirements.txt
# Django Core
Django==4.2.16
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Authentication & Security
django-cors-headers==4.3.1
argon2-cffi==23.1.0
django-oauth-toolkit==2.3.0
oauthlib==3.2.2

# API Documentation
drf-spectacular==0.27.2

# Filtering
django-filter==24.2

# Async Tasks
celery==5.4.0
redis==5.0.4

# Utilities
python-decouple==3.8
Pillow==10.3.0

# Server
gunicorn==22.0.0

# HTTP Client
requests==2.31.0

# Monitoring & Logging
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
